Restructured installer

+++ installer/includes/ximix/bin/run.sh
@@ -17,13 +17,19 @@ if [[ -z "$JAVA_HOME" ]]; then
 fi
 
 
-if [[ -z "$1" ]]; then
-        MIX="$XIMIX_HOME/conf/mixnet.xml";
+if [[ ! -z "$1" ]]; then
+        MIX="$XIMIX_HOME/$1/conf/mixnet.xml"
+	NODE="$XIMIX_HOME/$1/conf/node.xml"
 fi
 
-if [[ -z "$2" ]]; then
-        NODE="$XIMIX_HOME/conf/node.xml";
+if [[ ! -f "$MIX" ]]; then
+	echo "Network config not found for $1, path was $MIX";
+	exit -1
 fi
 
+if [[ ! -f "$NODE" ]]; then
+	echo "Node config was not found for $1, path was $NODE";
+	exit -1;
+fi
 
 $JAVA_HOME/bin/java -cp "$XIMIX_HOME/libs/*" org.cryptoworkshop.ximix.node.Main $MIX $NODE "$@"
